- reformat server side - move things around - add command fishingRep

+++ server/server.lua
@@ -1,4 +1,3 @@
-ps.versionCheck('md-fishing', 'https://raw.githubusercontent.com/Mustachedom/md-fishing/01f336f9cbbecb45dfe0d912887f1715663b3213/version.txt', 'https://github.com/Mustachedom/md-fishing/tree/01f336f9cbbecb45dfe0d912887f1715663b3213')
 
 ps.registerCallback('md-fishing:server:getLocations', function(source)
 	return locations
@@ -6,6 +5,7 @@ end)
 
 ps.registerCallback('md-fishing:server:getStores', function(source, type, location)
 	local src = source
+
 	if type == 'fishSales' then
 		if not ps.checkDistance(src, locations['fishBuyers'][location].loc, 5.0) then
 			ps.warn(ps.lang('Fails.tooFar', ps.getPlayerName(src), ps.lang('Info.fishBuyer')))
@@ -13,6 +13,7 @@ ps.registerCallback('md-fishing:server:getStores', function(source, type, locati
 		end
 		return stores.fishSales
 	end
+
 	if type == 'fishShop' then
 		if not ps.checkDistance(src, locations['shopLocation'][location].loc, 5.0) then
 			ps.warn(ps.lang('Fails.tooFar', ps.getPlayerName(src), ps.lang('Info.fishShop')))
@@ -25,10 +26,12 @@ end)
 
 RegisterNetEvent('md-fishing:server:sellFish', function(fish, loc)
 	local src = source
+
 	if not ps.checkDistance(src, locations['fishBuyers'][loc].loc, 5.0) then
 		ps.warn(ps.lang('Fails.tooFar', ps.getPlayerName(src), ps.lang('Info.fishBuyer')))
 		return
 	end
+
 	if fish == 'all' then
 		for k, v in pairs (stores.fishSales) do
 			local itemCount = ps.getItemCount(src, k)
@@ -41,7 +44,9 @@ RegisterNetEvent('md-fishing:server:sellFish', function(fish, loc)
 		end
 		return
 	end
+
 	if not stores.fishSales[fish] then return end
+
 	local itemCount = ps.getItemCount(src, fish)
 	if itemCount > 0 then
 		local total = itemCount * stores.fishSales[fish]
@@ -56,14 +61,17 @@ end)
 
 RegisterNetEvent('md-fishing:server:buyFishGear', function(loc, item, data)
 	local src = source
+
 	if not ps.checkDistance(src, locations['shopLocation'][loc].loc, 3.5) then
 		ps.warn(ps.lang('Fails.tooFar', ps.getPlayerName(src), ps.lang('Info.fishShop')))
 		return
 	end
+
 	if not stores['fishShop'][item] then
 		ps.warn(ps.lang('Fails.invalidItem', ps.getPlayerName(src), item, ps.lang('Info.fishShop')))
 		return
 	end
+
 	local price = stores['fishShop'][item] * data.amount
 	if ps.removeMoney(src, data.type, price) then
 		ps.addItem(src, item, data.amount)
@@ -75,23 +83,28 @@ end)
 
 ps.registerCallback('md-fishing:server:getBreakdown', function(source, location)
 	local src = source
+
 	if not ps.checkDistance(src, locations['breakDown'][location].loc, 3.5) then
 		ps.warn(ps.lang('Fails.tooFar', ps.getPlayerName(src), ps.lang('Info.breakDown')))
 		return
 	end
+
 	return breakDown
 end)
 
 RegisterNetEvent('md-fishing:server:breakDownRustys', function(loc, item)
 	local src = source
+
 	if not ps.checkDistance(src, locations['breakDown'][loc].loc, 3.0) then
 		ps.warn(ps.lang('Fails.tooFar', ps.getPlayerName(src), ps.lang('Info.breakDown')))
 		return
 	end
+
 	if not breakDown[item] then
 		ps.warn(ps.lang('Fails.invalidBreakdown', ps.getPlayerName(src), item, ps.lang('Info.breakDown')))
 		return
 	end
+
 	if ps.removeItem(src, item, 1) then
 		local itemReward = breakDown[item]['items'][math.random(1, #breakDown[item]['items'])]
 		local itemAmount = math.random(breakDown[item]['amount'].min, breakDown[item]['amount'].max)
@@ -99,6 +112,7 @@ RegisterNetEvent('md-fishing:server:breakDownRustys', function(loc, item)
 	end
 end)
 
+ps.versionCheck('md-fishing', 'https://raw.githubusercontent.com/Mustachedom/md-fishing/01f336f9cbbecb45dfe0d912887f1715663b3213/version.txt', 'https://github.com/Mustachedom/md-fishing/tree/01f336f9cbbecb45dfe0d912887f1715663b3213')
 
 RegisterNetEvent('md-fishing:server:inZone', function(locationType, location)
 	local src = source
@@ -114,6 +128,7 @@ RegisterNetEvent('md-fishing:server:outZone', function()
 	local src = source
 	local identifier = ps.getIdentifier(src)
 	inZonePlayers[identifier] = nil
+
 	if activeFishers[identifier] then
 		ps.notify(src, ps.lang('Fails.leftZone'), 'error')
 		activeFishers[identifier] = nil
@@ -132,15 +147,58 @@ local function getPoleType(pole)
 	return nil
 end
 
-local function getWaitTimer(identifier, poleType)
-	local levels = MySQL.query.await('SELECT levels FROM md_fishing WHERE citizenid = ?', { identifier })
-	if not levels[1] then return Config.Levels[0].time end
-	local lvl = json.decode(levels[1].levels)
-	if not lvl[poleType] then return Config.Levels[0].time end
-	return Config.Levels[lvl[poleType].level].time or Config.Levels[0].time
+local poles = {'fishingpole', 'magnetpole', 'illegalpole'}
+local spamProtection = {}
+
+local function checkSpam(license)
+	if spamProtection[license] then
+		return true
+	end
+	spamProtection[license] = true
+	CreateThread(function()
+		Wait(2000)
+		spamProtection[license] = nil
+	end)
+	return false
 end
 
-local poles = {'fishingpole', 'magnetpole', 'illegalpole'}
+local function checkSQL(src)
+	local identifier = ps.getIdentifier(src)
+	if cacheSQL[identifier] then
+		return true
+	end
+	local check = MySQL.query.await('SELECT * FROM md_fishing WHERE citizenid = ?', { identifier })
+	if not check[1] then
+		local levels = {
+			fishing = {level = 0, xp = 0},
+			illegal = {level = 0, xp = 0},
+			magnet = {level = 0, xp = 0},
+		}
+		MySQL.insert('INSERT INTO md_fishing (citizenid, levels, name) VALUES (?, ?,?)', {
+			identifier,json.encode(levels), ps.getPlayerName(src)
+    	})
+		cacheSQL[identifier] = levels
+	end
+	if not cacheSQL[identifier] then
+		cacheSQL[identifier] = json.decode(check[1].levels)
+	end
+	return true
+end
+
+ps.registerCallback('md-fishing:server:getLevels', function(source)
+	local src = source
+	local identifier = ps.getIdentifier(src)
+	checkSQL(src)
+	return {level = cacheSQL[identifier], maxXP = LevelUpAmount}
+end)
+
+local function getWaitTimer(identifier, poleType)
+	local lvl = cacheSQL[identifier]
+	if not lvl then return levelTime[0].time end
+	if not lvl[poleType] then return levelTime[0].time end
+	return levelTime[lvl[poleType].level].time or levelTime[0].time
+end
+
 for k, v in pairs (poles) do
 	ps.createUseable(v, function(source, item)
 		local src = source
@@ -152,6 +210,13 @@ for k, v in pairs (poles) do
 			TriggerClientEvent('md-fishing:client:stopfishing', src)
 			return
 		end
+
+		if checkSpam(identifier) then
+			return
+		end
+
+		checkSQL(src)
+
 		if not inZonePlayers[identifier] then
 			ps.notify(src, ps.lang('Fishing.notInZone'), 'error')
 			TriggerClientEvent('md-fishing:client:stopfishing', src)
@@ -164,6 +229,7 @@ for k, v in pairs (poles) do
 				TriggerClientEvent('md-fishing:client:stopfishing', src)
 				return
 			end
+
 			local checkDistance = #(GetEntityCoords(GetPlayerPed(src)) - locations.fishingZones[inZonePlayers[identifier].loc].loc)
 			if checkDistance > locations.fishingZones[inZonePlayers[identifier].loc].radius then
 				ps.notify(src, ps.lang('Fails.leftZone'), 'error')
@@ -171,16 +237,25 @@ for k, v in pairs (poles) do
 				TriggerClientEvent('md-fishing:client:stopfishing', src)
 				return
 			end
+
 			activeFishers[identifier] = {pole = v, location = inZonePlayers[identifier].loc, type = 'fishingZones', fishingType = getPoleType(v)}
-			TriggerClientEvent('md-fishing:client:fishing', src, getWaitTimer(identifier, getPoleType(v)))
+
+			timeouts[identifier] = getWaitTimer(identifier, getPoleType(v))
+			CreateThread(function()
+				Wait(1000 * timeouts[identifier])
+				timeouts[identifier] = nil
+			end)
+			TriggerClientEvent('md-fishing:client:fishing', src, timeouts[identifier])
 			return
 		end
-		if v == 'magnetpole' then
-			if inZonePlayers[identifier].type ~= 'fishingZones' then
+
+		if v == 'illegalpole' then
+			if inZonePlayers[identifier].type == 'fishingZones' then
 				ps.notify(src, ps.lang('Fishing.wrongPole'), 'error')
 				TriggerClientEvent('md-fishing:client:stopfishing', src)
 				return
 			end
+
 			local checkDistance = #(GetEntityCoords(GetPlayerPed(src)) - locations.illegalFishingZones[inZonePlayers[identifier].loc].loc)
 			if checkDistance > locations.illegalFishingZones[inZonePlayers[identifier].loc].radius then
 				ps.notify(src, ps.lang('Fails.leftZone'), 'error')
@@ -188,19 +263,22 @@ for k, v in pairs (poles) do
 				inZonePlayers[identifier] = nil
 				return
 			end
-			local levels = MySQL.query.await('SELECT levels FROM md_fishing WHERE citizenid = ?', { identifier })
-			if not levels[1] then
-				ps.notify(src, ps.lang('Fishing.noLevel'), 'error')
-				return
-			end
-			local lvl = json.decode(levels[1].levels)
+
+			local lvl = cacheSQL[identifier]
 			if lvl.illegal.level < StarIllLvl then
 				ps.notify(src, ps.lang('Fishing.wrongLevel', StarIllLvl), 'error')
 				TriggerClientEvent('md-fishing:client:stopfishing', src)
 				return
 			end
+
 			activeFishers[identifier] = {pole = v, location = inZonePlayers[identifier].loc, type = 'illegalFishingZones', fishingType = getPoleType(v)}
-			TriggerClientEvent('md-fishing:client:fishing', src,  getWaitTimer(identifier, getPoleType(v)))
+
+			timeouts[identifier] = getWaitTimer(identifier, getPoleType(v))
+			CreateThread(function()
+				Wait(1000 * timeouts[identifier])
+				timeouts[identifier] = nil
+			end)
+			TriggerClientEvent('md-fishing:client:fishing', src, timeouts[identifier])
 			return
 		end
 	end)
@@ -210,19 +288,22 @@ end
 --------------------------------- catch fish logic
 
 local function addRepToPlayer(identifier, poleType)
-	local levels = MySQL.query.await('SELECT levels FROM md_fishing WHERE citizenid = ?', { identifier })
-	if not levels[1] then return end
-	local lvl = json.decode(levels[1].levels)
+	local lvl = cacheSQL[identifier]
+	if not lvl then return end
 	if not lvl[poleType] then return end
+
 	lvl[poleType].xp = lvl[poleType].xp + 1
 	if lvl[poleType].xp >= LevelUpAmount then
-		lvl[poleType].level = lvl[poleType].level + 1
-		lvl[poleType].xp = 0
+		if lvl[poleType].level + 1 <= 10 then
+			lvl[poleType].level = lvl[poleType].level + 1
+			lvl[poleType].xp = 0
+		end
 	end
+
 	MySQL.update('UPDATE md_fishing SET levels = ? WHERE citizenid = ?', {json.encode(lvl), identifier})
 	CreateThread(function()
 		timeouts[identifier] = true
-		Wait(1000 * Config.Levels[lvl[poleType].level].time)
+		Wait(1000 * levelTime[lvl[poleType].level].time)
 		timeouts[identifier] = nil
 	end)
 end
@@ -230,21 +311,25 @@ end
 RegisterServerEvent('md-fishing:server:givefish', function()
 	local src = source
 	local identifier = ps.getIdentifier(src)
+
 	if timeouts[identifier] then
 		ps.notify(src, ps.lang('Fails.onTimeout'), 'error')
 		return
 	end
+
 	if not activeFishers[identifier] then
 		ps.notify(src, ps.lang('Fishing.notInZone'), 'error')
 		TriggerClientEvent('md-fishing:client:stopfishing', src)
 		return
 	end
+
 	if not inZonePlayers[identifier] then
 		ps.notify(src, ps.lang('Fails.leftZone'), 'error')
 		activeFishers[identifier] = nil
 		TriggerClientEvent('md-fishing:client:stopfishing', src)
 		return
 	end
+
 	local checkDistance = #(GetEntityCoords(GetPlayerPed(src)) - locations[activeFishers[identifier].type][activeFishers[identifier].location].loc)
 	if checkDistance > locations[activeFishers[identifier].type][activeFishers[identifier].location].radius then
 		ps.notify(src, ps.lang('Fails.leftZone'), 'error')
@@ -253,6 +338,7 @@ RegisterServerEvent('md-fishing:server:givefish', function()
 		TriggerClientEvent('md-fishing:client:stopfishing', src)
 		return
 	end
+
 	local bait = baitTypes[activeFishers[identifier].pole] or nil
 	if not bait then return end
 	local completed = false
@@ -266,6 +352,7 @@ RegisterServerEvent('md-fishing:server:givefish', function()
 			end
 		end
 	end
+
 	if not completed then
 		ps.notify(src, ps.lang('Fails.outOfBait'), 'error')
 		activeFishers[identifier] = nil
@@ -295,24 +382,12 @@ end)
 
 RegisterNetEvent('md-fishing:server:fishchum', function(location)
     local src = source
+
 	if not ps.checkDistance(src, locations['chumMaker'][location].loc, 3.5) then
 		ps.notify(src, ps.lang('Fails.tooFar', ps.getPlayerName(src), ps.lang('Info.breakDown')), 'error')
 		return
 	end
-	local chumMake = { -- % of the amount of fish will be turn to chum, example 100 fish = 90
-		flounder = 0.8,
-		reddrum = 0.8,
-		tunafish = 0.8,
-		steelhead = 0.8,
-		bluefish = 0.8,
-		halibut = 0.8,
-		catfish = 0.8,
-		whitebass = 0.8,
-		salmon = 0.8,
-		largemouthbass = 0.8,
-		panfish = 0.8,
-		trout = 0.8,
-	}
+
 	local amount = 0
     for k, v in pairs(chumMake) do
         local count = ps.getItemCount(src, k)
@@ -322,50 +397,29 @@ RegisterNetEvent('md-fishing:server:fishchum', function(location)
 			end
 		end
 	end
+
 	if amount <= 0 then
 		ps.notify(src, ps.lang('Shops.noFish', 'fish'), 'error')
 		return
 	end
 	ps.addItem(src, 'chum', amount)
 end)
+
  --- RUN THIS COMMAND BEFORE ANYONE JOINS THE SERVER. READ THE README FOR GODS SAKE THEN DELETE THIS COMMAND 
-ps.registerCommand('FishingReFormater', {
-	admin = true,
-}, function()
-	local allData = MySQL.query.await('SELECT * FROM mdfishing', {})
-	for k, v in pairs (allData) do
-		local levels = {
-			fishing = {level = v.fishing, xp = v.fishxp},
-			illegal = {level = v.illegalfish, xp = v.illfishxp},
-			magnet = {level = v.magnetfish, xp = v.magxp},
-		}
-		MySQL.update('INSERT INTO md_fishing (citizenid, levels, name) VALUES (?, ?, ?)', {v.citizenid, json.encode(levels), v.name})
-		MySQL.query.await('DELETE FROM mdfishing WHERE citizenid = ?', {v.citizenid})
-	end
-end)
 
-local function checkSQL(src)
-	local identifier = ps.getIdentifier(src)
-	local check = MySQL.query.await('SELECT citizenid FROM md_fishing WHERE citizenid = ?', { identifier })
-	if not check[1] then
-		MySQL.insert('INSERT INTO md_fishing (citizenid, levels, name) VALUES (?, ?,?)', {
-			identifier,json.encode({fishing = {level = 0, xp = 0}, illegal = {level = 0, xp = 0}, magnet = {level = 0, xp = 0}}), ps.getPlayerName(src)
-    	})
-	end
+if Config.OldData then
+	ps.registerCommand('FishingReFormater', {
+		admin = true,
+	}, function()
+		local allData = MySQL.query.await('SELECT * FROM mdfishing', {})
+		for k, v in pairs (allData) do
+			local levels = {
+				fishing = {level = v.fishing, xp = v.fishxp},
+				illegal = {level = v.illegalfish, xp = v.illfishxp},
+				magnet = {level = v.magnetfish, xp = v.magxp},
+			}
+			MySQL.update('INSERT INTO md_fishing (citizenid, levels, name) VALUES (?, ?, ?)', {v.citizenid, json.encode(levels), v.name})
+			MySQL.query.await('DELETE FROM mdfishing WHERE citizenid = ?', {v.citizenid})
+		end
+	end)
 end
-
-local function dropPlayer(src)
-	local identifier = ps.getIdentifier(src)
-	if activeFishers[identifier] then
-		activeFishers[identifier] = nil
-		TriggerClientEvent('md-fishing:client:stopfishing', src)
-	end
-	if inZonePlayers[identifier] then
-		inZonePlayers[identifier] = nil
-	end
-end
-
-RegisterNetEvent('QBCore:Server:OnPlayerLoaded', function() checkSQL(source) end)
-AddEventHandler('esx:playerLoaded', function(source) checkSQL(source) end)
-AddEventHandler('esx:playerDropped', function(source) dropPlayer(source) end)
-AddEventHandler('QBCore:Server:OnPlayerUnload', function(source) dropPlayer(source) end)
